[BC break] Change configuration loading in the extension
HTMLPurifier natively supports configuration inheritance. Since configuration profiles are already defined as services in the container, we can take advantage of that using the inherit() factory method to ensure that custom profiles all extend the "default" configuration..

This change affects BC since the merging strategy for configurations has changed. Profiles are no longer deep-merged, and it is now possible to disable caching defined by default for the "default" profile that other profiles extend.